Refactor main.rb and create new file lib/scraper.rb
- add Scraper class in scraper.rb and initialize it with url components
in main.rb
