+++ Gemfile
@@ -1,8 +1,8 @@
 source 'https://rubygems.org'
 
-gem 'sinatra'
+gem 'httparty'
 gem 'nokogiri'
 gem 'open-uri'
-gem 'httparty'
 gem 'rubocop', '~>0.81.0'
 gem 'sass'
+gem 'sinatra'
