+++ public/stylesheets/sass/main.scss
@@ -31,36 +31,38 @@ $enable-responsive-font-sizes: true;
                 font-size: 4rem;
                 text-align: center;
             }
-            form {
-                display: flex;
-                justify-content: center;
-                align-items: center;
-                position: relative;
-
-                .form-group {
-                    flex-grow: 1;
-                    height: 3rem;
-                    display: grid;
-                    place-items: center;
-                    
-                    #product_name {
-                        width: 100%;
-                        height: 100%;
-                        border-top-right-radius: 0;
-                        border-bottom-right-radius: 0;
-                    }   
-                }
-                .btn {
-                    border-top-left-radius: 0;
-                    border-bottom-left-radius: 0;
-                }
-
-                .bi-search {
-                    position: absolute;
-                    right: 8px;
-                }
-            }
         }
+    }
+}
+
+body .container form {
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    position: relative;
+
+    .form-group {
+        flex-grow: 1;
+        height: 3rem;
+        display: grid;
+        place-items: center;
+
+        #product_name {
+            width: 100%;
+            height: 100%;
+            border-top-right-radius: 0;
+            border-bottom-right-radius: 0;
+        }
+    }
+
+    .btn {
+        border-top-left-radius: 0;
+        border-bottom-left-radius: 0;
+    }
+
+    .bi-search {
+        position: absolute;
+        right: 8px;
     }
 }
